requirements/dev.txt
aiohttp==3.9.5 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
alembic==1.13.2 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.12" and python_version < "4.0"
asyncpg==0.29.0 ; python_version >= "3.12" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.12" and python_version < "4.0"
black==24.4.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.7.4 ; python_version >= "3.12" and python_version < "4.0"
cfgv==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
distlib==0.3.8 ; python_version >= "3.12" and python_version < "4.0"
dnspython==2.6.1 ; python_version >= "3.12" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi-cli==0.0.4 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.111.0 ; python_version >= "3.12" and python_version < "4.0"
filelock==3.13.4 ; python_version >= "3.12" and python_version < "4.0"
flake8==7.0.0 ; python_version >= "3.12" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.0.3 ; python_version < "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.12"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.12" and python_version < "4.0"
httptools==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.12" and python_version < "4.0"
identify==2.5.35 ; python_version >= "3.12" and python_version < "4.0"
idna==3.7 ; python_version >= "3.12" and python_version < "4.0"
isort==5.13.2 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
mako==1.3.5 ; python_version >= "3.12" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
mccabe==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.12" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.12" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
nodeenv==1.8.0 ; python_version >= "3.12" and python_version < "4.0"
orjson==3.10.6 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.12" and python_version < "4.0"
pathspec==0.12.1 ; python_version >= "3.12" and python_version < "4.0"
platformdirs==4.2.0 ; python_version >= "3.12" and python_version < "4.0"
pre-commit==3.7.0 ; python_version >= "3.12" and python_version < "4.0"
py-cord==2.5.0 ; python_version >= "3.12" and python_version < "4.0"
pycodestyle==2.11.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.20.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.3.4 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.8.2 ; python_version >= "3.12" and python_version < "4.0"
pyflakes==3.2.0 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.18.0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
python-multipart==0.0.9 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.12" and python_version < "4.0"
rich==13.7.1 ; python_version >= "3.12" and python_version < "4.0"
setuptools==69.5.1 ; python_version >= "3.12" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.31 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.37.2 ; python_version >= "3.12" and python_version < "4.0"
typer==0.12.3 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
ujson==5.10.0 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.30.1 ; python_version >= "3.12" and python_version < "4.0"
uvicorn[standard]==0.30.1 ; python_version >= "3.12" and python_version < "4.0"
uvloop==0.19.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.12" and python_version < "4.0"
virtualenv==20.25.3 ; python_version >= "3.12" and python_version < "4.0"
watchfiles==0.22.0 ; python_version >= "3.12" and python_version < "4.0"
websockets==12.0 ; python_version >= "3.12" and python_version < "4.0"
yarl==1.9.4 ; python_version >= "3.12" and python_version < "4.0"
